modify router shcema validator to allow multiple schema types
